GDM-86: make loading message be removed when WMS capabilities fails create abstract parent of both chooselayer steps
