+++ gwt/src/main/java/org/geomajas/plugin/deskmanager/client/gwt/manager/datalayer/steps/WmsChooseLayerStep.java
@@ -10,39 +10,29 @@
  */
 package org.geomajas.plugin.deskmanager.client.gwt.manager.datalayer.steps;
 
-import java.util.List;
-import java.util.Map;
-
-import com.google.web.bindery.event.shared.HandlerRegistration;
-import org.geomajas.gwt.client.Geomajas;
-import org.geomajas.plugin.deskmanager.client.gwt.manager.datalayer.DataReceiverWizardStep;
-import org.geomajas.plugin.deskmanager.client.gwt.manager.datalayer.NewLayerModelWizardWindow;
-import org.geomajas.plugin.deskmanager.client.gwt.manager.datalayer.Wizard;
-import org.geomajas.plugin.deskmanager.client.gwt.manager.datalayer.WizardStepPanel;
-import org.geomajas.plugin.deskmanager.client.gwt.manager.i18n.ManagerMessages;
-import org.geomajas.plugin.deskmanager.client.gwt.manager.service.DataCallback;
-import org.geomajas.plugin.deskmanager.client.gwt.manager.service.DiscoveryCommService;
-import org.geomajas.plugin.deskmanager.command.manager.dto.RasterCapabilitiesInfo;
-
 import com.google.gwt.core.client.GWT;
-import com.smartgwt.client.data.Record;
 import com.smartgwt.client.types.ListGridFieldType;
-import com.smartgwt.client.types.Overflow;
 import com.smartgwt.client.types.SelectionStyle;
-import com.smartgwt.client.widgets.Label;
 import com.smartgwt.client.widgets.grid.ListGrid;
 import com.smartgwt.client.widgets.grid.ListGridField;
 import com.smartgwt.client.widgets.grid.ListGridRecord;
 import com.smartgwt.client.widgets.grid.events.CellDoubleClickEvent;
 import com.smartgwt.client.widgets.grid.events.CellDoubleClickHandler;
-import com.smartgwt.client.widgets.grid.events.SelectionUpdatedEvent;
-import com.smartgwt.client.widgets.grid.events.SelectionUpdatedHandler;
+import org.geomajas.plugin.deskmanager.client.gwt.manager.datalayer.DataReceiverWizardStep;
+import org.geomajas.plugin.deskmanager.client.gwt.manager.datalayer.NewLayerModelWizardWindow;
+import org.geomajas.plugin.deskmanager.client.gwt.manager.datalayer.Wizard;
+import org.geomajas.plugin.deskmanager.client.gwt.manager.i18n.ManagerMessages;
+import org.geomajas.plugin.deskmanager.client.gwt.manager.service.DataCallback;
+import org.geomajas.plugin.deskmanager.client.gwt.manager.service.DiscoveryCommService;
+import org.geomajas.plugin.deskmanager.command.manager.dto.RasterCapabilitiesInfo;
+
+import java.util.List;
 
 /**
  * @author Jan De Moerloose
  * @author Jan Venstermans
  */
-public class WmsChooseLayerStep extends WizardStepPanel implements DataReceiverWizardStep {
+public class WmsChooseLayerStep extends AbstractChooseLayerStep implements DataReceiverWizardStep {
 
 	private static final ManagerMessages MESSAGES = GWT.create(ManagerMessages.class);
 
@@ -54,22 +44,21 @@ public class WmsChooseLayerStep extends WizardStepPanel implements DataReceiverW
 
 	private static final String FLD_INFO = "info";
 
-	private Map<String, String> connectionProps;
-
-	private ListGrid grid;
-
-	private Label warnings;
-
-	private String previousStep = NewLayerModelWizardWindow.STEP_WMS_PROPS;
-
-	private HandlerRegistration selectionUpdatedHandlerRegistration;
-
+	/**
+	 * Default constructor.
+	 *
+	 * @param parent
+	 */
 	public WmsChooseLayerStep(final Wizard parent) {
 		super(NewLayerModelWizardWindow.STEP_WMS_CHOOSE_LAYER, 
 			MESSAGES.wmsChooseLayerStepNumbering() + " "  + MESSAGES.wmsChooseLayerStepTitle(), 
-				false, parent);
+				false, parent, NewLayerModelWizardWindow.STEP_WMS_PROPS,
+				NewLayerModelWizardWindow.STEP_WMS_PREVIEW_LAYER);
 		setWindowTitle(MESSAGES.wmsChooseLayerStepTitle());
+	}
 
+	@Override
+	protected void createFieldsOfGrid() {
 		grid = new ListGrid();
 		grid.setWidth100();
 		grid.setHeight("*");
@@ -95,111 +84,38 @@ public class WmsChooseLayerStep extends WizardStepPanel implements DataReceiverW
 		descFld.setWidth("*");
 
 		grid.setFields(nameFld, crsFld, descFld);
-		grid.setCanResizeFields(true);
-
-		// -------------------------------------------------
-
-		warnings = new Label();
-		warnings.setWidth100();
-		warnings.setAutoHeight();
-		warnings.setPadding(3);
-		warnings.setOverflow(Overflow.VISIBLE);
-		warnings.setVisible(false);
-		warnings.setBackgroundColor("#FFCCCC");
-
-		addMember(grid);
-		addMember(warnings);
-	}
-
-	private void registerSelectionUpdateHandler(boolean register) {
-		if (register) {
-			selectionUpdatedHandlerRegistration = grid.addSelectionUpdatedHandler(new SelectionUpdatedHandler() {
-
-				public void onSelectionUpdated(SelectionUpdatedEvent event) {
-					fireChangedEvent();
-				}
-			});
-		} else  {
-			selectionUpdatedHandlerRegistration.removeHandler();
-		}
 	}
 
 	@Override
-	public void initialize() {
-		if (connectionProps != null) {
-			reset();
-			registerSelectionUpdateHandler(true);
-			grid.setShowEmptyMessage(true);
-			grid.setEmptyMessage("<i>" + MESSAGES.requestingInfoFromServer() +
-					" <img src='" + Geomajas.getIsomorphicDir() +
-					"/images/circle.gif' style='height: 1em' /></i>");
-			
-			grid.redraw();
-			DiscoveryCommService.getRasterCapabilities(connectionProps,
-					new DataCallback<List<RasterCapabilitiesInfo>>() {
+	protected void executeCapabilitiesStep() {
+		DiscoveryCommService.getRasterCapabilities(connectionProps,
+				new DataCallback<List<RasterCapabilitiesInfo>>() {
 
-						public void execute(List<RasterCapabilitiesInfo> result) {
-							//First add all map crs
-							for (RasterCapabilitiesInfo wli : result) {
-								ListGridRecord lgr = new ListGridRecord();
-								lgr.setAttribute(FLD_NAME, wli.getName());
-								lgr.setAttribute(FLD_CRS, wli.getCrs());
-								lgr.setAttribute(FLD_DESC, wli.getDescription());
-								lgr.setAttribute(FLD_INFO, wli);
-								grid.addData(lgr);
-							}
-							//First add all other crs
+					public void execute(List<RasterCapabilitiesInfo> result) {
+						//First add all map crs
+						for (RasterCapabilitiesInfo wli : result) {
+							ListGridRecord lgr = new ListGridRecord();
+							lgr.setAttribute(FLD_NAME, wli.getName());
+							lgr.setAttribute(FLD_CRS, wli.getCrs());
+							lgr.setAttribute(FLD_DESC, wli.getDescription());
+							lgr.setAttribute(FLD_INFO, wli);
+							grid.addData(lgr);
 						}
-					}, new DataCallback<String>() {
+						//First add all other crs
+					}
+				}, new DataCallback<String>() {
 
-						public void execute(String result) {
-							reset();
-							warnings.setVisible(true);
-							warnings.setContents("<b><i>" + result + "</i></b>");
-						}
-					});
-		}
-	}
-
-	@Override
-	public void setData(Map<String, String> connectionProps) {
-		this.connectionProps = connectionProps;
-	}
-
-	@Override
-	public boolean isValid() {
-		return (grid.getSelectedRecord() != null);
-	}
-
-	@Override
-	public String getNextStep() {
-		return NewLayerModelWizardWindow.STEP_WMS_PREVIEW_LAYER;
-	}
-
-	@Override
-	public String getPreviousStep() {
-		return previousStep;
-	}
-
-	public void setPreviousStep(String previousStep) {
-		this.previousStep = previousStep;
-	}
-
-	@Override
-	public void reset() {
-		grid.deselectAllRecords();
-		grid.setData(new Record[] {});
-		warnings.setVisible(false);
-		warnings.setContents("");
+					public void execute(String result) {
+						showFailMessage(result);
+					}
+				});
 	}
 
 	@Override
 	public boolean stepFinished() {
-		final WmsPreviewLayerStep nextStep = (WmsPreviewLayerStep) parent
-				.getStep(NewLayerModelWizardWindow.STEP_WMS_PREVIEW_LAYER);
+		final WmsPreviewLayerStep nextStep = (WmsPreviewLayerStep) parent.getStep(getNextStep());
 		RasterCapabilitiesInfo info = (RasterCapabilitiesInfo) grid.getSelectedRecord().getAttributeAsObject(FLD_INFO);
 		nextStep.setData(connectionProps, info);
-		registerSelectionUpdateHandler(false);
-		return true;
+		return super.stepFinished();
 	}
 }
